Changing the timing for the special effects.
When you hit the wall, we draw things.
If we change that every frame it's hard to see.
We were updating it every 100ms.
Now we update the screen after a random amount of time.
The time varies from 100ms to 175ms.
It looks more like a hand drawn cartoon now.

Added then commented out some code.
Code to draw a purple splat on the near wall of the cube.
Drawing the wall itself would never work.
Adding the splat didn't help, either.