Slow the frame but only for the ball.
Cut the ball speed in half, because that looked better.
It really needed to be faster with the original frame rate.

We're still tracking the ball at the full framerate.
And we always draw the ball the instant it touches a wall.

This is part of a bigger change, described in README.md.
I stopped here because it looked nice and because it was interesting.
You can really see and appreciate the 𝑟𝑜𝑢𝑔ℎ appearance of the ball.
Especially when it hits the near wall and goes "splat" and gets "smashed".
The ball is not blurry at all.
But the ball is jumpy.
README.md suggests a few ways to deal with that.