Add controls and fix sound.
Add controls.
By default eveything is muted.
You have to click to unmute.
Without this chrome gets upset and the program doesn't have permission to makes sounds.

In the previous commit sound was basically broken.
Except, temporarily, in my test setup!

Added a speed control.
It doesn't do anything yet.
It was eaiser to design the GUI for all the controls at once.