Change an icon to look better on the mac.
The unicode white question mark❔ has a thick black border on the PC.
It has a thin gray border on the Mac.
It looked like it was disabled.

The unicode "black" question mark ❓ is red on both systems.
It works, so I'm using it now.

I also added notes about a sound problem on Mac's Safari browser.