Improve the control panel.
You cannot select the text from the control panel any more.
I did that on accident sometimes and it looked bad.

You can now click on the 🐢 or the 🐰 to change speed.
The cursor changes to show if the button icon is available or disabled.

Fix a TODO item.
The initial velocity now uses the same algorithm
as when the user changes the velocity.